Rename and move dummy_input to connector.dummy.input
