+++ logprep/connector/connector_factory.py
@@ -8,7 +8,7 @@ from logprep.connector.connector_factory_error import (
 )
 from logprep.input.input import Input
 from logprep.output.output import Output
-from logprep.input.dummy_input import DummyInput
+from logprep.connector.dummy.input import DummyInput
 from logprep.input.jsonl_input import JsonlInput
 from logprep.input.json_input import JsonInput
 from logprep.connector.confluent_kafka.input import (
