+++ tests/unit/connector/test_connector_factory.py
@@ -19,7 +19,7 @@ from logprep.connector.connector_factory import (
 from logprep.connector.confluent_kafka.input import (
     ConfluentKafkaInput,
 )
-from logprep.input.dummy_input import DummyInput
+from logprep.connector.dummy.input import DummyInput
 from logprep.input.json_input import JsonInput
 from logprep.input.jsonl_input import JsonlInput
 from logprep.connector.confluent_kafka.output import (
